+++ NEWS.md
@@ -2,8 +2,10 @@
 - released: ???
 - added `permute()`
 - improved filtering by axioms specified in `excluded` argument of `dig*()`
+- improved the core search algorithm by better handling of condition prefixes
+  and caching of conjunction supports
 - added `"search_stats"` attribute to each result of `dig*()` functions
-- improved documentation
+- improved documentation, added vignettes
 
 # nuggets 2.2.2
 - released: 2026-07-22